# Deflation rewrite