--- a/src/NEPTypes.jl
+++ b/src/NEPTypes.jl
@@ -1220,6 +1220,7 @@ Returns true/false if the NEP is sparse (if compute_Mder() returns sparse)
 
 
     include("nep_transformations.jl")
+    include("nep_deflation.jl")
     include("low_rank_nep.jl")
 
     # structure exploitation for DEP