This fixes #166 and

* Makes deflation more user friendly

* Makes deflation more efficient, in paricular for sparse matrices.

* Makes deflation of a deflated problem not a recursive deflated problem, but just a modified.

* Introduces three deflation modes: Generic, SPMF, MM. SPMF can be used if the SPMF is short and gives a AbstractSPMF. It has some drawbacks in terms of robustness. Generic seems to be the most efficient. 

Some timing:

```julia
nep=nep_gallery("dep0_sparse",20);
n=size(nep,1);
local λ,v;
(λ,v)=quasinewton(nep,v=ones(n),λ=3,tol=1e-11,displaylevel=1,maxit=200)
```
```julia
dnep1=deflate_eigpair(nep,λ,v,mode=:Generic);
@btime quasinewton(dnep1,v=ones(size(dnep1,1)),λ=1+1im,tol=1e-11)
```
  2.638 ms (16661 allocations: 3.28 MiB)
```julia
dnep2=deflate_eigpair(nep,λ,v,mode=:SPMF);
@btime quasinewton(dnep2,v=ones(size(dnep1,1)),λ=1+1im,tol=1e-11)
```
  10.774 ms (54923 allocations: 5.88 MiB)

Old approach:
```
dnep3=deflate_eigpair(nep,λ,v,mode=:MM);
@btime quasinewton(dnep3,v=ones(size(dnep1,1)),λ=1+1im,tol=1e-11)
```
  6.127 ms (24827 allocations: 4.76 MiB)



If we change to method involving the higher derivatives with `compute_Mder`, we
obtain:
```julia
nep=nep_gallery("dep0_sparse",100);
n=size(nep,1);
(λ,v)=quasinewton(nep,v=ones(n),λ=2,tol=1e-11,displaylevel=1,maxit=200)
```
```julia
dnep1=deflate_eigpair(nep,λ,v,mode=:Generic);
@btime mslp(dnep1,λ=1im,tol=1e-11)
```
  44.163 ms (3188 allocations: 18.30 MiB)
```julia
dnep2=deflate_eigpair(nep,λ,v,mode=:SPMF);
@btime mslp(dnep2,λ=1im,tol=1e-11)
```
  43.594 ms (4020 allocations: 12.08 MiB)
Old approach:
```julia
dnep3=deflate_eigpair(nep,λ,v,mode=:MM);
@btime mslp(dnep3,λ=1im,tol=1e-11)
```
  426.791 ms (2500 allocations: 101.48 MiB)



How to use it to find several eigenvalues:

```julia
  nep=nep_gallery("dep0");
  n=size(nep,1);
  local λ,v;
  (λ,v)=quasinewton(nep,v=ones(n),λ=0,tol=1e-11) # Warmup
  local dnep
  dnep=deflate_eigpair(nep,λ,v,mode=:Generic);
  for k=1:4
      (λ,v)=augnewton(dnep,v=ones(size(dnep,1)),λ=-1+0.1im,
                      tol=1e-11,displaylevel=1,maxit=300,armijo_factor=0.5)
      dnep=deflate_eigpair(dnep,λ,v); # Deflate the computed eigval
  end
  (λv,V)=get_deflated_eigpairs(dnep) # Extract eigpair information
  for k=1:size(λv,1)
     @show norm(compute_Mlincomb(nep,λv[k],V[:,k])) # Prints essentially zeros
  end
```