@eringh 2019-03-08
Potentially we want to move the function `jd_inner_effenberger_linear_solver!` from JD to the deflation. That function uses the `LinSolver` and a Schur complement to solve the associated linear system. What do you think?