@jarlebring 2019-03-08
Yes. That code could be useful also outside the context of JD. I view it as a Schur complement linsolver, which is quite natural for the deflated problems. Maybe it should even go to `LinSolvers`. This PR is so big already, and the current setup is not completely wrong, so I would prefer waiting a bit with transferring that code to somewhere more general.