@eringh 2019-03-08
Jag har också tittat på en fix för JD. Tror att med den nya strukturen så kan JD utnyttja den bättre.