@jarlebring 2019-03-08
Excellent. Just let me merge this PR (a couple of small things remaining).