@eringh 2019-03-08
Ok. Then I'll put the JD improvements in a separate PR after this is merged.